controller: datastore: fix race with deleteInstance and AddInstance
There is a window of opportunity for the tenant cache to be
empty for a tenant that is having instances launched for it
if a start error is received at the right time.  Prevent this
by holding a lock on the tenant cache while refreshing the
cache data and deleting the tenant usage.

Signed-off-by: Kristen Carlson Accardi <kristen@linux.intel.com>