ciao-launcher: docker: Fix a bug in docker networking
Fix a bug in which docker containers could not be created if a qemu
instance had been previously created on the same node with the same
subnet.  Launcher was assuming that as the subnet had already been
created that the docker network had also been created, but this
is of course not always the case.  The fix works by always attempting
to create the docker network the first time launcher is asked to
create a docker network for a new subnet.

There's a possibility that launcher could try to create a docker
network that already exists, if it is restarted after the network
was originally created.  In this case we simply ignore the erorr.
If the network failed to be created and did not previously exist
an error will be returned when the container is subsequently created.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>