ciao-controller: compute: Add total number of instances
We add the total number of instances that one would get when
not filtering/restricting the response.
For example if 150 instances are running for a given tenant, but
we only ask the first 10, we'd get the first 10 instances and
total_servers would be 150.

This is an enhancement over the standard compute API that allows
UIs to do an efficient paginations. They can compute the total
number of pages while only displaying one of them.

Signed-off-by: Samuel Ortiz <sameo@linux.intel.com>