ciao-launcher: hard-reset: Delete all ciao docker networks
This commit fixes

https://github.com/01org/ciao/issues/32

It works by deleting all docker networks that it can find, on the
current node, that use the ciao network plugin.  Previously we only
deleted the networks used by the instances that had just been deleted.
However, we found that networks could be leaked under certain
circumstances, if for example, launcher crashed or the node shutdown
while performing a hard-reset.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>