datastore: move sql into it's own file
Separate the sql code from the general datastore code more
cleanly by moving it into it's own file and just having it
be exposed as an interface.  One day we might wind up needing
to try different sql databases, and this design allows one
to easily replace the sqlite3 queries with a different backend.
It also makes it more clear when you are going to incur the
cost of going to the database vs. just using cached data.

Signed-off-by: Kristen Carlson Accardi <kristen@linux.intel.com>