controller: workload csv typo asks for 80vcpus for docker
This is probably not a real issue today, but...

Each of our sample workloads appears meant to request 80MB of persistent
storage.  Which is ignored anyway.  There is a typo for two docker workload
descriptions which sets the resource_id to #2 for the disk line instead of

Resource #2 is cpu, so technically the template is saying those docker
workloads should get 80 CPUs worth of resource allocation.  That's not
good.  Now really we don't account CPUs or create cpusets (that I'm aware
of anyway) yet for docker workloads.  And it may be that having read the
earlier line requesting 2 CPUs worth of resource, that the second request
is ignored.

So we're probably multiple times lucky.  Either way, let's just make
sure we don't rely on luck.