ciao-launcher: Set simulation ticker to nil once expired
We close the connected channel when the ticker expires. If we
get 2 really close ticks they will be queued and we will end
up closing the connected channel twice.
To prevent that from happening we just set it to nil after the
first tick and it thus is removed from the select fd set.

Signed-off-by: Samuel Ortiz <sameo@linux.intel.com>