3rd party: Add missing license for cancellable
This license is missing in the version of engine-api that we use so the vendoring
tool doesn't pick it up.  Still we don't want to be in violation of the license
of this piece of code so I'm adding it manually.  The vendoring tool will pick
up the license when we move to the next version of engine-api.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>