ssntp: Add InvalidConfiguration error type
The InvalidConfiguration error is either
sent by the Scheduler to report an invalid
CONFIGURE payload back to the sender, or by
the clients to which a CONFIGURE command
has been forwarded to and that leads to
configuration errors on their side.
When the scheduler receives such error back
from any client it should revert back to the
previous valid configuration.

Signed-off-by: Samuel Ortiz <sameo@linux.intel.com>