ciao-scheduler: remove old test code
These are left-overs from the very first week of the project when there was
nothing to talk to the initial scheduler ssntp server.  I'm going to
swizzle them into a scheduler_test.go, but in the meantime 'go vet'
complains they're unreachable.  They're probably unbuildable even without
patching.  Better to just remove them and replace with a proper test.

Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>