travis: Fail build on go vet, golint or misspell errors
go vet is run on all packages except for ciao-controller which currently
contains a go vet error.  Once this is fixed we'll update the travis.yml
file to vet this package as well.

The following packages are excluded from golint runs as they too contain
errors:

- github.com/01org/ciao/ciao-cli
- github.com/01org/ciao/ciao-controller/internal/datastore
- github.com/01org/ciao/ciao-controller/types

golint is disabled on tip for the time being.  It seems to be panicking.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>