ciao-scheduler: fix deadlock in connect/disconnect notifiers
Commit a17d38a2 introduced a data deadlock between connect and disconnect
notifiers.  The connect notifier has a role, but disconnect does not.  This
means the disconnect routine acquires and defers the unlock of multiple
maps' protection mutex.

We should probably have the disconnect include a role id, so we only
acquire the one needed lock.  In the meantime:

The maps are read mostly so have a RW lock to protect the map structure.
We take the reader lock in paths which simply traverse the map.  We take
the write lock in paths we know are going to modify the map itself.
Map member objects have their own locks if needed to protect the object's
internal data consistency.  Using a read lock while traversing the map to
forward notifications fixes side of the issue.

And not doing deferred unlocks fixes the other side, though it is better to
use defers if we can add a role id to the disconnect notifier.  Issue #121
opened to track this.

Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>