Networking: CN API: Split Create VNIC into multiple functions
Split create VNIC into multiple functions to reduce the overall
complexity. One function handles the case where the bridge and
tunnel already exists (the more common case) and the other
handles the case where the bridge, tunnel and vnic all have
to be created.

The critical section gets carried over into the callee function.
This is needed for performance reasons as we want to keep the
high latency device creation logic outside the critical section.

Signed-off-by: Manohar Castelino <manohar.r.castelino@intel.com>