ciao-scheduler: yet more lock fixing
I apparently mismerged/misrebased a change.  The earlier code was supposed
to be leaving nodeStat's locked once picked for dispatching.  I left some
Unlock() calls in the code.  So the later code that's supposed to unlock
the locked node was unlocking an unlocked node, which nicely results in a
very obvious panic from golang.

Signed-off-by: Tim Pepper <timothy.c.pepper@intel.com>