Networking: Bug Fix: Loopback Interfaces should be excluded
Currently the Travis CI flag is also used to build the all
in one test setup. The loopback interface was being considered
a valid physical interface causing problems on the CNCI.

Loopback devices are now explicitly excluded from being used
even in the travis and all in one setup.

A dummy interface with a non routable IP should be used for
travis and unit testing.

Longer term we need seperate flag --allinone vs using the
travis tag for all in one setup

Signed-off-by: Manohar Castelino <manohar.r.castelino@intel.com>