test-cases: Create coverage profile
test-cases now has support for generating a coverage profile for the
tests that it executes.  This behaviour is controlled by two new
command line options, --coverprofile and --append-profile.  When
specified by itself, --coverprofile will generate a new coverage file
with the data from each package tested.  When used with --append-profile,
the coverage data generated by test-cases will be appended to an
existing coverage file.  You should only append to coverage files
generated by test-cases as test-cases assumes that the coverage mode is
always 'set', which it may not be if the file was generated by another
tool.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>