ciao-controller: Select the right certificates
With issues/67 fixed, we no longer can use a random
certificate as the client role will be defined by
the client certificate OIDs.

By default we are now using cert-Controller-localhost.pem
and cert-CNAgent-localhost.pem when the client wants to
run as an AGENT. Unit tests fail to pass without that fix.

Signed-off-by: Samuel Ortiz <sameo@linux.intel.com>