documentation: push vs pull
The scheduler is a pull scheduler more than a push scheduler, or now I'm
hearing people say "push/pull".  The scheduler does not simply push work,
looking for a place to land it.  Compute nodes check in and request work,
with the scheduler handing the work to a node that has requested work.
It's not strictly pull arguably because the compute nodes aren't checking
in and immediately bidding for or pulling work off a work queue.

Signed-off-by: Tim Pepper <timothy.c.pepper@intel.com>