ssntp: Automatically assign SSNTP server role
We now parse the server certificate and look for
a set of OIDs. We then translate that into an SSNTP
roles bitmask and force the server into using that
role.
The server will fail to listen for new connections
if the requested role through config does not match
the client certificate.

This will allow us to automatically detect which role
a server should use based on its certificate and not expect
servers to explictely specify a role.

Fixes #67

Signed-off-by: Samuel Ortiz <sameo@linux.intel.com>