ciao-launcher: Abstract SSNTP connection.
This commit adds an interface to abstract the underlying SSNTP connection.
This will be useful in unit testing as it will allow us to test the
launcher code that sends data to SSNTP servers, without actually having
a real server.

The launcher code has been updated to send SSNTP command, errors and
events via this interface and not directly via the real SSNTP connection.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>