Networking: CNCI Agent: Enable multiqueue support
Enable multiqueue support within the CNCI using the CNCI systemd
init script. This script today sets the number of queues to 4
which is the default value configured by the launcher.

In the future the clear linux with auto detect the number of
queues on the device and seup the queues to match.

This allows the CNCI to support full 10Gbps line rate.

The ethtool command is allows to fail to handle the case where
the CNCI is running on top of VM in the single machine developer
setup.

Note: If the agent is restarted manually within the CNCI this
will also fail even on hardware as ethtool returns error even
when the queue setup matches what is desired. This is the case
when the agent is restarted manually as it ran on boot.

This commit closes https://github.com/01org/ciao/issues/95

Signed-off-by: Manohar Castelino <manohar.r.castelino@intel.com>