ciao-scheduler: remove useless old code
The node picker had an old special case for when there's just one
compute node.  IIRC that was there from a point when my newbie golang
skills were outputing code that puked trying to traverse a larger than
size one set of nodes.  It was never useful though because you're not
truly performance critical if you've only got one node.  And besides
you're still going to find the one node pretty darned fast even if it's
not as fast as straight referencing the zero'th element in a list.

Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>