ssntp: make Role a more opaque type
The Role is a typedef'd bitmask of an enumeration of constant values.
But much of the code directly accessed the underlying integer, with a
variety of casting involved.  By removing all the integer typing and
adding some accessor/test functions, the casts go away and and some code
reads more succintly.  We get a more strongly typed object for "Role" that
can transparently change its underlying implementation in the future if
needed.

Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>