client: ClusterConfiguration() returns a formatted payload
We can't expect callers to know that a configuration blob
maps to a payloads.Configure structure so we have to provide it
from the SSNTP client API instead.

NOTE: The actual YAML unmarshalling will be done from the
configuration package as the logic and internal configuration
knowledge (data structure, blob to structure mapping) belongs
there.

Signed-off-by: Samuel Ortiz <sameo@linux.intel.com>