ciao-controller: Set compute URL to machine name
For single machine setup the compute URL should be automatically
set to the hostname. This allows the CNCI to connect to the
scheduler running on the host (i.e. the hostname).

The compute URL was previosly set to localhost. However this
causes problems to the CNCI as the launcher will send local
host as the scheduler address to the CNCI. This means that the
scheduler will try connect to local host (within the CNCI
which will not work)

Signed-off-by: Manohar Castelino <manohar.r.castelino@intel.com>