ssntp: Add a sync channel to the SSNTP configuration
This is an optional field that callers can use if they
need to get notified about their Dial() or Serve() call.

Typically SSNTP servers may want to launch a go routine
to handle SSNTP clients. By creating an error channel and
adding it to their config they will be able to listen
on this channel to wait for Serve() to complete.

Signed-off-by: Samuel Ortiz <sameo@linux.intel.com>