ciao-vendor: Add uses command.
The uses command allows users to determine which other packages
depend upon a specific package.  The list includes both direct
and indirect dependencies.  This command is useful when
removing or unvendoring a component.  It allows you to figure out
how the removal of a package would affect ciao as a whole.

Fixes #192

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>