testutil: add integrated clients and server test coverage
This commit represents the bulk of the testutil/ test coverage.
It implements a set of integrated client/server tests using the testutil
agent, controller and server.

Controller gains a set of channel functions like those in agent and server.
For all three the result getter for events has a 20s timeout to account for
the reconnect pause in clients when the server goes away, which happens in
TestReconnects().

Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>