test-cases: Output stderr from test cases after failure
Test-cases now gathers up all the output it receives on stderr
when running go test.  If any of the packages under test fail,
test-cases will dump out the collected stderr when it exits.
This should make it easy to figure out what went wrong.  Right
now all you know is that a test failed, which is not enough
information.

Fixed #184

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>