ciao-launcher: Split ovsCh into two channels.
Using one ovsCh for delivering commands to the overseer from both
the main and the instance go routine was racy.  There was always
the chance that an instance routine could write to the ovsCh after
it had been closed by the main go routine but before the instance
had itself been asked to shutdown by the overseer, which would
lead to a panic.

Fixes #26

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>