testutil: agent SendStats() reimplementation
Again races and inconsistencies in the implementation.

A test STATS payload factory function and constants are introduced.
This allows common usage by the payloads/stats_test.go (for a "const"
struct) and caio-controller (which uses a randomized set of instances).

A caller that cares about the actual result needs to run the agent
SendStats() as a go routine and await the channel response.  The
ciao-controller/compute_test.go does this, but for reasons unknown to
me the ciao-controller/controller_test.go was not accounting the return.
I've not changed that.

Additionally it's worth noting ciao-controller/compute_test.go has a
lot of one or two second sleeps.  I optimistically removed those after
adding the channel around the SendStats() call, but got inconsistent
test results.  I didn't get to the bottom of why the sleeps magically
tend to allow the tests to pass.

Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>