testutils: Single VM CI: Reduce CNCI Resources
The single VM test setup can be used to launch a small number
of workloads (VMs and Containers). In this setup the resources
required by the CNCI are significantly less. Hence reduce
the CNCI profile to be minimal. This will reduce the dev VM's
memory requirements significantly.

This fixes issue https://github.com/01org/ciao/issues/233

Signed-off-by: Manohar Castelino <manohar.r.castelino@intel.com>