ciao-cnci-agent: Obtain the scheduler address from cert
SSNTP certificates contain the addresses of valid schedulers
for a cluster. This information can be used directly by the
CNCI agent to connect to the scheduler.

This commit eliminates the need for the the launcher to send
in the scheduler FQDN/IP through the ciao config drive.

Also as this was the only parameter needed by the the CNCI agent
for ciao the config drive is no longer needed.

The CNCI agent still allows the explicit specification of the
the Scheduler URI. If the URI is explicitly specified then the
URI is used first (in addition to the URI's in the certs).

This fixes issue https://github.com/01org/ciao/issues/236

Signed-off-by: Manohar Castelino <manohar.r.castelino@intel.com>