release: add ciao BAT tests
Add an "_release" directory to the ciao project. The "_" prefix
tells the go compiler to ignore this directory. This directory
will contain code related to the ciao release process. This commit
adds a "bat" subdirectory under the _release directory with a single
python test script. This test script is documented with python
docstring in the code so that pydoc can be used to extract the
latest documentation. A README file is also included for
convenience, which contains the output of pydoc.

Signed-off-by: Kristen Carlson Accardi <kristen@linux.intel.com>