ciao-vendor: Add revendor command
ciao-vendor now supports a revendor command that allows one to
update a given repository to a new version.  When revendoring,
all the packages in the repository that are used by ciao are
necessarily updated.  An example of its use is:

go run ciao-vendor/ciao-vendor.go revendor github.com/tylerb/graceful 510b0c6

Fixes #70

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>