testutil: add channel test coverage
The error/event/command results channels were mostly tested but
implicitly in the client-server test.  Explicit coverage means they can be
thoroughly tested.

Because the timeouts are tested, this does slow the test run down.
Expecially for the 20s timeout error and event result channels, it might
make sense to condition them to only run when not doing a short test.

Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>