Add compatibility for Compute API /v2.1/{tenant_id}/servers POST
Add compatibility for OpenStack with CIAO Compute API call /v2.1/{tenant_id}/servers
for POST method . CIAO was missing to return the 'server' variable which
should include the details of the first server built, now we return same data
as CIAO expects with 'servers' and 'totalServers' plus 'server'.
CIAO internally uses the status 'running' but OpenStack uses 'active',
changing the value to 'active' when a API call is made.
For a complete API reference please refer to
http://developer.openstack.org/api-ref-compute-v2.1.html .
Fixes #209

Signed-off-by: Leoswaldo Macias <leoswaldo.macias@intel.com>