ciao-controller: Add proper http return codes to showServerDetails
The function showServerDetails attending calls from
/v2.1/{tenant}/servers/{server} GET Method was returning 500
InternalServerError for all error cases, this fix covers
Unauthorized, NotFound and InternalServer return  errors for each
corresponding case. To accomplish this we made next:
1. Created the function 'returnErrorCode' to avoid code and call
it with the corresponding values where needed.
2. Also we are adding two structures (HTTPReturnCodeData and
HTTPReturnCode) which will be marshalled to return RESPONSE BODY
for the call as OpenStack does, but adding a generic "error".
http://developer.openstack.org/api-guide/compute/faults.html

Fixes #158

Signed-off-by: Leoswaldo Macias <leoswaldo.macias@intel.com>