testutil: add ssntp.Status test Result channels
Agents send ssntp.Status changes for server consumption.  As with the
event/error/command testutil.Result channels, we need one for status
messages to be tracked in test cases to verify delivery.

A next patch will introduce usage of the new status channel in a broader
test case, but here at least is unit coverage of the helpers in isolation.

Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>