test-cases: Output logs of failed tests to stderr
Test-cases now gathers up the logs of failed test cases and outputs them
after the report.  Logs are grouped by unit test so it should be easy
to locate the cause of the failure.  Also, we only output logs from failed
test-cases so these logs are not hidden amongst the logs of those that
succeed.  Failed logs are currently only output for textual reports.

Fixes #295

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>