networking: Remove graceful HTTP server from the docker plugin
This commit replaces the use of the graceful HTTP server with the HTTP server
provided by the Go standard library.  After careful consideration it was
determined that ciao does not currently need the functionality provided by
this library.  The graceful HTTP server tries to shut down all open connections
before shutting down the HTTP server itself.  However, we do not need to do
this in ciao as we can be sure that the docker plugin's HTTP server
can have no outstanding connections when it is asked to shutdown by
ciao-launcher.

Fixes #29
Fixes #221

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>