ciao-launcher: refactor instance go routine code
The code for the instance go routine is split over a number of
different files.  These files have been renamed to make it
obvious in which context their code runs.  Code that is not
unique to the instance go routine, i.e., the error types for
specific commands has been moved to new files.

Fixes: #10

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>