testutil: TestStart wait on server response before agent
The testutil client-server test's TestStart function waits first on the
agent and then on the server.  That's technically fine since this is all
asynchonous and threaded, but it's illogical.  The command first flows
through the server so wait on seeing it flow through the server first.
Then wait on seeing the command at the agent.

Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>