test-cases: Add timeout and verbose options
test-cases can now be configured to output the name of the package it is
about to test to stdout.  This is important as doing so will reset the
Travis CI kill timer.  If this is not done, test-cases has only 10 minutes
to run all the unit tests for all the specified packages.  Specifying
the verbose option gives you 10 minutes per package.

test-cases can also be configured to timeout the unit tests for a package
after a specified amount of minutes.  It will then proceed to run the
unit tests on any remaining packages, if any, before outputting the results.
Timeout packages are killed via a SIGABRT.P

Fixes #328

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>