ciao-launcher: Update the -network parameter
-network is now a bool parameter and it defaults to true.  Previously, it
was an emumerated type that accepted none, cn, nn, or dual.  But these
additional values are no longer needed as SSNTP can deduce the role
from its certificates.  Another important note is that -network defaults
to true.  Previously, if the networking flag was unspecified, networking
was disabled.  This default behaviour was left over from the early days
of ciao when we were integrating networking.  Networking on by default
is now the sensible choice.

Fixes #162

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>