release: add release notes generator to ciao tree
ciao-release is a small program which will generate the release
notes for ciao releases. It uses the github api to obtain
the list of pull requests closed on the master branch since
the time of the last tag, up to the release candidate sha. The rc
sha must be passed in as a string argument. It also gets the list of issue
events since the last tag and reports any closed issues.
Finally, it gets the list of all commits since the last tag.
The lists are output into a plain text file that is called
"release.txt". This file is used to create an annotated tag
for the next release.

Signed-off-by: Kristen Carlson Accardi <kristen@linux.intel.com>