ciao-launcher: Limit the number of parallel starts
This commit limits the number of parallel starts to a function
of the number of CPUs present in the node.  There really isn't
much point in allowing 1000 instances to be started on the same
node at the same time.  Doing so won't increase the start times
much and will increase the likelihood of failure due to the
resource exhaustion caused by the heavy demands of instance
startup.

Fixes #8

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>