ciao-controller: add more synchonization points to test
The ciao-controller tests have mostly relied on time.Sleep() for "ordering"
of events.  At least for the portions which flow through a testutil client
or server, it's possible to use the event and command channels to see that
the expected events have traversed the stack in the expected places.

The sleep's can't fully, simply be removed though because the controller
does not have a similar mechanism to insure its asynchronous processing has
completed before carrying on.

Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>