testutil: ssntp.ConcentratorInstanceAddedEvent event flows
Add ssntp.ConcentratorInstanceAddedEvent event flows to the testutil
client-server test.  This means adding a cnciAgent client and a test
case which simulates the controller requesting via the server that a
netAgent start the cnciAgent into the testutil client-server test.

Similarly ciao-controller's tests which did a less thorough simulation of
the ssntp.ConcentratorInstanceAddedEvent event flow need adjusted.

Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>