ciao-controller: make returnErrorCode func variadic
returnErrorCode function receives a message string  which it will return
with this design the message needed to be formatted before calling the
function, making the function variadic for the message will enable the
function to be call with the messages to be formatted in the function,
so we don't have to preformat every time we call it.

The returnErrorCode func will now build the message string in the format
specified with the args passed to it as values.

Signed-off-by: Leoswaldo Macias <leoswaldo.macias@intel.com>