ssntp: add stringer tests
The test coverage of these stringer functions was about 50% as they happen
to be called in some places by some tests.  Easy to make it 100% and then
if somebody changes the strings it highlights that output may be parsed by
things which would break as the result of such changes.

Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>