ciao-controller: datastore: handle volume stats
When an instance reports that a volume is attached via stats,
make sure that we have this attachment stored. If the attachment
disappears (indicating the volume detached) then make sure we
delete the attachment. When a detach/attach is noted, make sure
to update the block data to keep the state correct.

Signed-off-by: Kristen Carlson Accardi <kristen@linux.intel.com>