ciao-controller: enable testHTTPRequest to send invalid token
The function testHTTPRequest is used to send the http requests, and
it always sends requests with valid tokens, we are enabling it to
send invalid tokens for negative test cases to fail in authentication
by not sending the token in the request if the validToken variable is
false. Now all calls to testHTTPRequest need to specify if the should
have a valid token or not.

Fixes #349

Signed-off-by: Leoswaldo Macias <leoswaldo.macias@intel.com>