ciao-controller: convert compute test cases into templates
Compute tests is only covering happy paths, we are converting all
testcases into templates, this way we can re-use their code to cover
also negative cases, such as invalid authentication and other scenarios.
With this change we use the template to avoid duplicating code when
covering same functionality.

Also this commit contains the required changes for the testcases that cover
the happy path to use its corresponding template.

The idea of the template is that we can reuse it and only call with different
parameters, we might need to change the templates for coming test integrations
but would be minimal changes and would keep the code cleaner that if we
start duplicating code.

Templates created:

 - testCreateServer
 - testListServerDetailsTenant
 - testListServerDetailsWorkload
 - testShowServerDetails
 - testDeleteServer
 - testServersActionStart
 - testServerActionStop
 - testListFlavors
 - testShowFlavorDetails
 - testListFlavorsDetails
 - testListTenantResources
 - testListTenantQuotas
 - testListEventsTenant
 - testListNodeServers
 - testListTenants
 - testListNodes
 - testNodeSummary
 - testListCNCIs
 - testListCNCIDetails
 - testListTraces
 - testListEvents
 - testClearEvents
 - testTraceData

Fixes #349

Signed-off-by: Leoswaldo Macias <leoswaldo.macias@intel.com>