ciao-controller: add negative test cases for validation of invalid auth
ciao-controller/compute_test.go was missing in its test coverage  for all the
scenarios for of failed authentication.
These changes increases the code coverage for the previous common scenario. It
was required to create new test cases which call the templates of its
functionality but with invalid parameters for authentication and also needed to
modify their templates so to stop evaluating the template because the following
instructions would not apply to them if the invalid token scenario was provided.

The new created negative test cases are:

 - TestCreateSingleServerInvalidToken
 - TestListServerDetailsTenantInvalidToken
 - TestShowServerDetailsInvalidToken
 - TestDeleteServerInvalidToken
 - TestServersActionStartInvalidToken
 - TestServerActionStopInvalidToken
 - TestListFlavorsInvalidToken
 - TestShowFlavorDetailsInvalidToken
 - TestListFlavorsDetailsInvalidToken
 - TestListTenantResourcesInvalidToken
 - TestListTenantQuotasInvalidToken
 - TestListNodeServersInvalidToken
 - TestListTenantsInvalidToken
 - TestListNodesInvalidToken
 - TestNodeSummaryInvalidToken
 - TestListCNCIsInvalidToken
 - TestListCNCIDetailsInvalidToken
 - TestListTracesInvalidToken
 - TestListEventsInvalidToken
 - TestClearEventsInvalidToken
 - TestTraceDataInvalidToken

 Fixes #349

Signed-off-by: Leoswaldo Macias <leoswaldo.macias@intel.com>