ciao-controller: add negative testcase for invalid tenantServersAction
The function tenantServersAction has multiple scenarios (os-start,
os-stop, os-delete) for multiple actions, but also verifies the
requested action is valid, if not fail, this commit contains the required
changes to to verify that, if the action is not valid, then verify it.
To make this, we created the template testServersActionStop which receives
two extra parameters, the expected http status and the action to perform

In order to align the new code we had to made the TestServersActionStop
to call it with the action  "os-start" and its corresponding http status,
and then we added a new test cases, TestServersActionStopWrongAction,
which sends an invalid action, "wrong-action", with the correspoding http
status (StatusServiceUnavailable).

Fixes #349

Signed-off-by: Leoswaldo Macias <leoswaldo.macias@intel.com>