testutil: delete channel after closing it
The code to put a result on a channel checks to see that something has
posted as awaiting a result for the given command/event/error/status
before writing to the channel to avoid writing to a non-existing channel.
That can still write to a closed channel if the closed channel is not
removed from the map.

Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>