testutil: rework results channel locking
The locking on the testutil results channels was incorrect.  The channel
writers block and do so with a lock held.  That's a major no-no.  Defering
a lock unlock seems nice a clean...unlock noted immediately after a lock.
But then you might block and sleep.

To straighten this out, the locks need initialized once at the time the
ssntp test actor is created.  Only the channel maps get modified inside the
channel close/open helpers, and those only do such modificiations with the
appropriate lock held.

Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>