ciao-cli: Handle the case where a user does not belong to a tenant.
When a tenant is not specified in the cli or environment variables,
ciao-cli uses getTenant function to fetch the tenantName and
tenantID the user belongs to.

getTenant was not handling the case where a user exists but does not
belong to any tenant resulting in an index out of range panic.

This commits handles this scenario and fails properly.

Fixes: #433

Signed-off-by: Alberto Murillo <alberto.murillo.silva@intel.com>