ciao-launcher: Ensure qmp ids are valid ids
The commands to attach and detach a volume currently fail if the
UUID of that volume does not begin with a letter.  The UUIDs are
used in QMP device and drive names and the QMP commands to which
these names are passed can fail, if the UUIDs start with a number.
This patch fixes the issue by prepending, rather than appending,
drive_ or device_ to the volume UUIDs before that are used in QMP
ids.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>