ciao-controller: add basic headers to method for definition of API Docs
CIAO's APIs are growing and will keep growing while new features are
added to its functionality. Nowdays we in a early stage and need to have
every functionality docummented, as CIAO keeps growing it will be harder
to maintain the documentation up to date.

To prevent human errors we are adding the automation of API docs, this
will be done by using the tool https://github.com/yvasiyarov/swagger,
which aims us to automate the process of API documentation following
a "Declarative Comments Format".

This commit is adding:
 - Main API documentation
 - SubApi definition
 - For all methods define:
	- Descriptions
 	- Call
 	- HTTP verbs
 	- Success results
 	- Failure Results

Signed-off-by: Leoswaldo Macias <lmaciasm10@gmail.com>